JPA 기본편, 활용 2 적용 완료

+++ src/main/java/com/hsj/force/domain/entity/TIngredient.java
@@ -1,12 +1,15 @@
 package com.hsj.force.domain.entity;
 
+import com.hsj.force.domain.entity.embedded.CommonData;
 import com.hsj.force.domain.entity.embedded.TIngredientId;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
@@ -19,7 +22,7 @@ public class TIngredient {
     private TIngredientId ingredientId;
 
     private String ingredientName;
-    private int quantity;
+    private double quantity;
     private String insertId;
     private LocalDateTime insertDate;
     private String modifyId;
